requirements.txt
audiocraft==1.3.0
diffusers==0.30.3
ipython==8.12.3
Pillow==10.4.0
PySide6==6.7.3
Requests==2.32.3
torch==2.4.1+cu124
transformers==4.45.1
accelerate==0.26.0